requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
build==0.10.0
    # via pip-tools
click==8.1.4
    # via pip-tools
colorama==0.4.6
    # via
    #   build
    #   click
contourpy==1.1.0
    # via matplotlib
cycler==0.11.0
    # via matplotlib
fonttools==4.40.0
    # via matplotlib
future==0.18.3
    # via -r requirements.in
h5py==3.9.0
    # via mat73
kiwisolver==1.4.4
    # via matplotlib
mat73==0.60
    # via -r requirements.in
matplotlib==3.7.2
    # via
    #   -r requirements.in
    #   seaborn
numpy==1.25.1
    # via
    #   contourpy
    #   h5py
    #   mat73
    #   matplotlib
    #   pandas
    #   scipy
    #   seaborn
packaging==23.1
    # via
    #   build
    #   matplotlib
pandas==2.0.3
    # via
    #   -r requirements.in
    #   seaborn
pillow==10.0.0
    # via matplotlib
pip-tools==6.14.0
    # via -r requirements.in
pyparsing==3.0.9
    # via matplotlib
pyproject-hooks==1.0.0
    # via build
python-dateutil==2.8.2
    # via
    #   matplotlib
    #   pandas
pytz==2023.3
    # via pandas
scipy==1.11.1
    # via -r requirements.in
seaborn==0.12.2
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
tzdata==2023.3
    # via pandas
wheel==0.40.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
